backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.10.6
pydantic-settings==2.7.1
langchain==0.1.0
langchain-openai==0.0.6
chromadb>=1.5.7
sentence-transformers==2.7.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx>=0.27.0
openai>=1.10.0
pyyaml==6.0.1
pandas==2.1.4
tqdm==4.66.1
redis==5.2.1
psycopg[binary]==3.2.3
spacy==3.8.2
bcrypt>=4.1.2
PyJWT>=2.8.0